+++ vr_exclude_back.h
@@ -5,7 +5,7 @@
 
 typedef struct Vr_Exclude_Back_ Vr_Exclude_Back;
 struct Vr_Exclude_Back_ {
-   uint32_t hash;
+   uint64_t hash;
    Int nbBack;
    Addr ip[BACKTRACE_SIZE];
    Bool used;
@@ -32,7 +32,7 @@ struct Vr_Back_ {
    VgFile * handlerAddrList;
 };
 
-uint32_t hash_back(Int nbBack, Addr* ip);
+uint64_t hash_back(Int nbBack, Addr* ip);
 
 void vr_back_init(Vr_Back* vrBack, Bool genExclude, const HChar* rep);
 void vr_back_finalize(Vr_Back* vrBack);
